% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_get_group}
\alias{resourcegroups_get_group}
\title{Returns information about a specified resource group}
\usage{
resourcegroups_get_group(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the resource group.}
}
\description{
Returns information about a specified resource group.
}
\section{Request syntax}{
\preformatted{resourcegroups$get_group(
  GroupName = "string"
)
}
}

\keyword{internal}
