% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resourcegroups_operations.R
\name{resourcegroups_update_group_query}
\alias{resourcegroups_update_group_query}
\title{Updates the resource query of a group}
\usage{
resourcegroups_update_group_query(GroupName, ResourceQuery)
}
\arguments{
\item{GroupName}{[required] The name of the resource group for which you want to edit the query.}

\item{ResourceQuery}{[required] The resource query that determines which AWS resources are members of
the resource group.}
}
\description{
Updates the resource query of a group.
}
\section{Request syntax}{
\preformatted{resourcegroups$update_group_query(
  GroupName = "string",
  ResourceQuery = list(
    Type = "TAG_FILTERS_1_0"|"CLOUDFORMATION_STACK_1_0",
    Query = "string"
  )
)
}
}

\keyword{internal}
